Remove the game folder
