+++ src/engine/unix_assets.cpp
@@ -14,7 +14,6 @@
 
 #include "header.h"
 
-#include "../game/game_includes.h"
 #include "util/debug.cpp"
 #include "math/block_math.h"
 #include "asset/asset.h"
